Update README.md
The combined project aims to conduct a comprehensive analysis utilizing data from two Power BI dashboards: one focusing on an e-commerce online store's performance, and the other providing insights into COVID-19 statistics. The objective is to derive actionable insights, trends, and forecasts to enhance business strategies for the e-commerce platform while contributing to understanding the dynamics and impact of the COVID-19 pandemic.